fix up CLI tools for standalone mapping generation
