+++ mapping-env/bin/ipython
@@ -1,0 +1,8 @@
+#!/mnt/dev/TISLAB/monarch-gene-mapping/mapping-env/bin/python
+# -*- coding: utf-8 -*-
+import re
+import sys
+from IPython import start_ipython
+if __name__ == '__main__':
+    sys.argv[0] = re.sub(r'(-script\.pyw|\.exe)?$', '', sys.argv[0])
+    sys.exit(start_ipython())
