+++ monarch_gene_mapping/download.yaml
@@ -1,0 +1,37 @@
+---
+# Alliance
+-
+  url: https://fms.alliancegenome.org/download/BGI_FB.json.gz
+  local_name: data/alliance/BGI_FB.json.gz
+  tag: alliance_gene
+-
+  url: https://fms.alliancegenome.org/download/BGI_MGI.json.gz
+  local_name: data/alliance/BGI_MGI.json.gz
+  tag: alliance_gene
+-
+  url: https://fms.alliancegenome.org/download/BGI_RGD.json.gz
+  local_name: data/alliance/BGI_RGD.json.gz
+  tag: alliance_gene
+-
+  url: https://fms.alliancegenome.org/download/BGI_SGD.json.gz
+  local_name: data/alliance/BGI_SGD.json.gz
+  tag: alliance_gene
+-
+  url: https://fms.alliancegenome.org/download/BGI_WB.json.gz
+  local_name: data/alliance/BGI_WB.json.gz
+  tag: alliance_gene
+-
+  # url: https://fms.alliancegenome.org/download/BGI_XB.json.gz
+  url: https://download.alliancegenome.org/5.1.1/BGI/XB/1.0.1.4_BGI_XB_1.json.gz
+  local_name: data/alliance/BGI_XB.json.gz
+  tag: alliance_gene
+-
+  url: https://fms.alliancegenome.org/download/BGI_ZFIN.json.gz
+  local_name: data/alliance/BGI_ZFIN.json.gz
+  tag: alliance_gene
+
+# hgnc
+-
+  url: http://ftp.ebi.ac.uk/pub/databases/genenames/hgnc/tsv/hgnc_complete_set.txt
+  local_name: data/hgnc/hgnc_complete_set.txt
+  tag: hgnc_gene
